api_gateway/requirements.txt
flask>=3.1.1,<3.2
flask_sqlalchemy>=3.1.1,<3.2
flask_migrate>=4.0.5,<4.1

psycopg2-binary>=2.9.3,<2.10
authlib>=1.3.0,<1.4

python-dotenv>=1.0.1,<1.1
requests>=2.31.0,<3.0
